style: reformat code for better readability
